washr: Publication Toolkit for Water, Sanitation and Hygiene (WASH) Data

A toolkit to set up an R data package in a consistent structure. Automates tasks like tidy data export, data dictionary documentation, README and website creation, and citation management.

See also

Author

Maintainer: Lars Schöbitz lschoebitz@ethz.ch (ORCID)

Authors:

Lars Schöbitz lschoebitz@ethz.ch (ORCID)

Mian Zhong mzhong@ethz.ch (ORCID)

Margaux Götschmann margauxg@ethz.ch (ORCID)

Colin Walder cwalder@ethz.ch (ORCID)

Other contributors:

Global Health Engineering, ETH Zurich [copyright holder]